Generalized enumArray to enumIndexedSeq.
This retains the enumArray function for backwards compatibility,
but it now just delegates to enumIndexedSeq. In theory, there's
nothing functionally that prevents this from being generalized
further to take a Seq instead, except that the implementation
would prefer to have a performant apply(). Since you'd likely
only want to feed this an indexed sequence, preferring a
different enumeration for a non-indexed one, it seems safer to
only generalize it this far.